rate機能

+++ db/migrate/20220304044503_create_master_pieces.rb
@@ -6,7 +6,7 @@ class CreateMasterPieces < ActiveRecord::Migration[6.1]
       t.string :jenre
       t.text :introduction
       t.integer :user_id
-      t.integer :rate
+      t.float :rate
       t.timestamps
     end
   end
